Switch from random to UCT-based exploration (worse performance)
Switched from completely random exploration to prefering further
exploration of the most promising nodes, based on estimated value
so far and the uncertainty associated with the number of times
that they had already been explored. I was surprised that Figurer's
overall performance dropped slightly. I have no explanation for
that drop and feel that Figurer needs better metrics to help me
investigate such issues. I am saving this commit in a side branch
with the intention of improving Figurer's metrics and coming back
to this issue later.