Create simple 'Robot 2D' example to test figurer
Created Robot 2D example to serve as a simple test case for
figurer, and likely to also serve as a 'hello world' level
example for explaining how to use figurer in the documentation.
Expanded figurer_distribution to include a uniform distribution.
Added tests for uniform distribution and, when those tests
found a bug in the Distribution class, fixed that bug. Added
failing tests based on Robot 2D, which fail because
Context::figure_once is not yet implemented.