Calculate error bars for node values
Calculated error bars for node values because error estimates
will be helpful later for choosing exploration direction and
for plan sampling. Exploration follows the child node with
the highest upper bound on value. The final plan follows the
child with the highest lower bound on value.

Calculated error components because the largest error component
suggest how the node should be refined next. If error is caused
primarily by sparsity (not enough child nodes) then the next step
should be to create another child node. If the error is caused
primarily by child errors, then one of the child nodes should be
further refined.

Sparsity error for nodes with at least two children is based on
the value spread for those children. The idea is to estimate how
much the value would change if a child node were added at the
high or low end of the existing child values, maintaining the
prior average distance between child values. When a node has
only one child, its future value spread is assumed to be similar
to the root node.

The natural approach to calculating child-based error for state
nodes is to use highest lower bound out of all child nodes and
the highest upper bound out of all child nodes as the upper and
lower bound of this node. Unfortunately, that approach leads
to severe fluctuations in the upper bound as new child nodes
with high error are added. The approach is modified slightly
to reduce the effect of one outlier on this calculation.