Prevent positive feedback loop between value and sparsity error
Sparsity error for a state node was limited to the positive
direction because a new child node with higher value would be
selected to raise the state node's value, while a new child
node with lower value would just be ignored as irrelevant.
While this approach makes theoretical sense, it creates a
destabilizing positive feedback loop. Large sparsity errors
increase the expected value of a state node, leading to
higher values of newly created root children, leading to
higher value spread between older and newer root children,
leading to higher default value for sparsity error,
completing the loop to further raise the value of new nodes.

The solution was simply to pretend that sparsity error is
two-tailed, such that sparsity error does not shift the
expected value for a node. The most likely negative effect
of this change would be to discourage poorly explored nodes
from being selected for the final plan. That negative effect
seems acceptable because such nodes are unlikely to be
selected for other reasons.