Create distribution node that aims for existing state node
Used predict_inverse_fn_ to select actuation that will likely
result in connection to a specific other state node. This
approach substantially increases the cross-connections between
different iterations.

The final results on Robo2D test are sometimes very good, such as
this recent plan whose final position was only 0.4 away from
the goal of (3,4).

1.00, 2.00
0.80, 0.04  =>  1.79, 2.04
0.01, 0.00  =>  1.80, 2.04
0.70, 0.77  =>  2.50, 2.81
0.69, 0.41  =>  3.19, 3.21
0.16, 0.63  =>  3.36, 3.83

However, good results like above are not achieved consistently.
Often the last step, especially, will have an unreasonable
move away from the goal.

The Robo2D test finishes 100 iterations in just 38 ms, which is
surprisingly fast considering that it relies heavily on a
naive implementation of spatial_index that will need to be
optimized later.

Likely next steps:
1) Encourage spreading on a global level, reaching states that
are far from what any node has reached in the past.
2) Encourage spreading on a local level, reaching states that
are far from what any other direct child of this node has
reached.
3) Encourage further exploration of lower depth nodes.
Inter-connections lead to the possibility that the same node
will be a successor of two nodes with different depths, and will
itself have a depth that is only adequate for one of them.